Updated tests, added a few more, prints are more clear
