+++ scs2-examples/src/main/resources/urdf/torso.urdf
@@ -1,5 +1,5 @@
 <?xml version="1.0" ?>
-<robot name="TorsoWithArm">
+<robot name="torso">
     <link name="baseLink">
         <visual>
             <geometry>
