+++ scs2-examples/src/main/resources/urdf/torsoAndLimb.urdf
@@ -1,0 +1,79 @@
+<?xml version="1.0" ?>
+<robot name="torsoAndLimb" xmlns:xacro="http://ros.org/wiki/xacro">
+    <link name="baseLink">
+        <visual>
+            <geometry>
+                <sphere radius="0.1"/>
+            </geometry>
+            <material>
+                <color rgba="0.2 0.7 0.9 1"/>
+            </material>
+            <origin rpy="0 0 0" xyz="0.1 0 0.3"/>
+        </visual>
+    </link>
+    <joint name="jointA" type="revolute">
+        <axis xyz="0 1 0"/>
+        <origin rpy="0 0 0" xyz="0.1 0 0.2"/>
+        <parent link="baseLink"/>
+        <child link="bodyAB"/>
+        <limit effort="190" lower="-1.1" upper="0.4141" velocity="5.89"/>
+    </joint>
+    <link name="bodyAB">
+        <inertial>
+            <mass value="0.025"/>
+            <origin rpy="0 0 0" xyz="0  0  0.1"/>
+            <inertia ixx="1.0e-8" ixy="0" ixz="0" iyy="1.0e-8" iyz="0" izz="1.0e-8"/>
+        </inertial>
+        <visual>
+            <geometry>
+                <cylinder radius="0.0125" length="0.4"/>
+            </geometry>
+            <material>
+                <color rgba="0.92 0.85 0.2 1"/>
+            </material>
+            <origin rpy="0 0 0" xyz="0  0  0.1"/>
+        </visual>
+    </link>
+    <joint name="jointB" type="revolute">
+        <axis xyz="0 1 0"/>
+        <origin rpy="0 0 0" xyz="0 0 0.2"/>
+        <parent link="bodyAB"/>
+        <child link="endEffectorLink"/>
+        <limit effort="190" lower="-1.1" upper="0.4141" velocity="5.89"/>
+        <mimic joint="jointA" multiplier="1.25" offset="0"/>
+    </joint>
+    <link name="endEffectorLink">
+        <inertial>
+            <mass value="5"/>
+            <origin rpy="0 0 0" xyz="0  0  0.1"/>
+            <inertia ixx="0.01" ixy="0" ixz="0" iyy="0.01" iyz="0" izz="0.005"/>
+        </inertial>
+        <visual>
+            <geometry>
+                <box size="0.0125 0.0125 0.2"/>
+            </geometry>
+            <material>
+                <color rgba="0.2 0.7 0.9 1"/>
+            </material>
+            <origin rpy="0 0 0" xyz="0 0 0.1"/>
+        </visual>
+    </link>
+    <joint name="jointC" type="revolute">
+        <axis xyz="1 0 0"/>
+        <origin rpy="0 0 0" xyz="0.1 0 0.0"/>
+        <parent link="baseLink"/>
+        <child link="head"/>
+        <limit effort="190" lower="-1.1" upper="0.4141" velocity="5.89"/>
+    </joint>
+    <link name="head">
+        <visual>
+            <geometry>
+                <sphere radius="0.1"/>
+            </geometry>
+            <material>
+                <color rgba="0.2 0.7 0.9 1"/>
+            </material>
+            <origin rpy="0 0 0" xyz="0.1 0 0.3"/>
+        </visual>
+    </link>
+</robot>
