+++ scs2-examples/src/test/java/us/ihmc/scs2/examples/urdf/ComposedURDFLoadingTest.java
@@ -1,0 +1,68 @@
+package us.ihmc.scs2.examples.urdf;
+
+import org.junit.jupiter.api.Test;
+import us.ihmc.scs2.definition.robot.RobotDefinition;
+import us.ihmc.scs2.definition.robot.urdf.URDFTools;
+import us.ihmc.scs2.definition.robot.urdf.items.URDFModel;
+
+import javax.xml.bind.JAXBException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * Class that tests the ability to load a urdf that is split up in multiple files, this may be don't because the urdf
+ * is rather large or has parts that are often replaced.
+ */
+public class ComposedURDFLoadingTest
+{
+   @Test
+   public void loadURDFFromOneFiles() throws JAXBException
+   {  // Ensure loading a single urdf file works
+      String[] urdfFiles = {"urdf/torsoAndLimb.urdf"};
+      URDFModel urdfModel = getURDFModel(urdfFiles);
+
+      URDFTools.URDFParserProperties parserProperties = new URDFTools.URDFParserProperties();
+      RobotDefinition robotDefinition = URDFTools.toRobotDefinition(urdfModel, parserProperties);
+      assertEquals(3, robotDefinition.getAllJoints().size());
+   }
+
+   @Test
+   public void loadURDFFromTwoFiles() throws JAXBException
+   {
+      String[] urdfFiles = {"urdf/torso.urdf", "urdf/limbA.urdf"};
+      URDFModel urdfModel = getURDFModel(urdfFiles);
+
+      URDFTools.URDFParserProperties parserProperties = new URDFTools.URDFParserProperties();
+      RobotDefinition robotDefinition = URDFTools.toRobotDefinition(urdfModel, parserProperties);
+      // We only expect 3 because we don't keep any duplicates within the files
+      assertEquals(3, robotDefinition.getAllJoints().size());
+   }
+
+   @Test
+   public void loadURDFFromThreeFiles() throws JAXBException
+   {
+      String[] urdfFiles = {"urdf/torso.urdf", "urdf/limbA.urdf", "urdf/limbB.urdf"};
+      URDFModel urdfModel = getURDFModel(urdfFiles);
+
+      URDFTools.URDFParserProperties parserProperties = new URDFTools.URDFParserProperties();
+      RobotDefinition robotDefinition = URDFTools.toRobotDefinition(urdfModel, parserProperties);
+      // We only expect 3 because we don't keep any duplicates within the files
+      assertEquals(3, robotDefinition.getAllJoints().size());
+   }
+
+   private static URDFModel getURDFModel(String[] urdfFilenames) throws JAXBException
+   {
+      Collection<InputStream> inputStreams = new ArrayList<>();
+
+      for (String urdfFile : urdfFilenames)
+      {
+         inputStreams.add(ComposedURDFLoadingTest.class.getClassLoader().getResourceAsStream(urdfFile));
+      }
+
+      return URDFTools.loadURDFModel(inputStreams, Collections.emptyList(), SimpleCrossFourBarURDFRobot.class.getClassLoader());
+   }
+}
